Accumulate animation blend weights along the animation tree
When an animation tree is processed, the parent's blend weights are
ignored when computing a child node's blend weights. This leads to
incorrect final blend weights, reporting 1.0 for animations that actually
don't contribute at all.

With this commit, blend weights are passed down the animation tree so
that the values read from the animation states match the actual blend
weights.